fix problem checking server_info
Fix problem with server_info exposed in  PullRequest #5
The problem was that the method called from wrong entity. It's a method of MongoClient, not of a collection. (programming without internet and wtihout mongo.. something happens XDDD) I cannot test it right now I havent the hardware and i wont to install a mongo right now, but it should be fix.

Co-Authored-By: Dídac García <didaquis@users.noreply.github.com>